Update - v1.1.1.1
Added

- Added black list for all modules
Now you can set the just the modules you want for your conversation(group or private)

- Added debug file
In debug folder, will be created the log files(in development, not debuging all things yet)

- Added base modules
All the modules have a base called BotModulesCore. In this base, all that is equal in every modules, was abstracted in one base parent to fill all the children modules.

- Added new module - MediaBot
You can(using google as back) search on the media sites to retrieve some song, profile or video.
Right know it's working on YouTube, YouTube Music, Spotify, Soundcloud and Twitter(more will be added in the future)

- Added new modules - SupplyBot
You can take a list(that's is just what you wanna know, i promisse to you)

- Added a database called credentials.xml
There you will put your credentials of google search, news and climate API

- Added a print based function
To print in the console what is log, core or other types of debug, was created a function to do help with this.
Now it's just a bunch of methods that have the new print function, but in the future will replace all prints in the project.

- Added .gitignore to the repository

Modified

- Moved media/images folder to the root of app
- Moved profile/wpp folder to the root of app
- Config file will be created if doesn't exists
- All the events, when expired, it's saved in another database called roles_done.xml
- Watson bot disabled temporarily
- Updated ChatterBot(1.0.5) and it's dependencies
- Help module now can retrieve just the help of some specific module