Update - v1.1.1.2
Changelog v1.1.1.2

Added

- Added new module - GameBot:
Now you can play old lady game with your friends with domi!!
The module have a ranking and a match counts, to work with this datas together in the future.
There a chess game, but not working yet, just the board can be viewed with 'domi xadrez --board'.

- Added new base class - Functions:
The BotModulesClass is extended by function, where inherit all definitions the is base of the whole application to pass to the modules
including CoreBot that also inherit from it.

- Added a 'module', that is not a module right know because it's locate on the soul of bot - InterMessages:
With this, you can send messages to another contact or group with Domi by group or private. You can send with a
name that will display your number or just with --anon parameter, that will not show information of who is sending the message.

- Added a function to mention someone in a group by number

Modified
- Improved the way that Domi listen to the messages
It also retrieve the label name(not real name, just the name that the user choose on WhatsApp) and the number to work with some modules and tasks
- The conversation now has the last_message unique by its conv.
- Some databases files will be created automatically in case that is not exist
- The main bot was splited into two pieces:
bot_start.py will just create a instance of the bot_main.py, initialize and mantain the bot_main updating the inputs in real-time
bot_main.py will hold the inputs of the application(the inputs sended by the users on WhatsApp and collected by Selenium)
- Removed linux chromedriver of the folder

Removed
- Twitter bot(for the moment)

Futures modifications and appends:
- More improve on way that Domi listen to the messages
Work with some way, that it will listen all the past messages, in case of messages sended to it when it is on another conversation,
To mention the messages directed to it and to pick context of the conversations, improving not the answers but the 'brain' of the bot.
- Improve of the machine learning of the responses, create a full algorithm to that.
- Save the large and sensitive data in MySql or another related database, with encriptation, to improve otimization, security of the informations and performance.

* Fix of bugs and otimizations